Update esg.cmip6.ini
Set the default for the "create CIM" behaviour to true for CMIP6, as discussed on a recent PWT telco.

This variable will be honoured by the code which has been committed to https://github.com/IS-ENES-Data/esg-publisher/tree/cim-integration-2, and which is the subject of a pull request at https://github.com/IS-ENES-Data/esg-publisher/pull/3, which I have asked for people at IS-ENES-Data team to review, hopefully merge into IS-ENES-Data/esg-publisher/devel and then create an onward pull request into ESGF/esg-publisher/devel.  If that pull request comes your way, then this one should be merged to accompany it.